move faerie-related code from function_manifest to output
...and since `function_manifest` is no longer needed after this, delete it.
